3392: Merge mobile and desktop header language selectors
